Move register page store to central auth store
