--- src/containers/RegisterPage/constants.js
@@ -1,5 +1,0 @@
-export const REGISTER = '[Register]';
-
-export const REGISTER_REQUEST = `${REGISTER} REGISTER_REQUEST`;
-export const REGISTER_SUCCESS = `${REGISTER} REGISTER_SUCCESS`;
-export const REGISTER_ERROR = `${REGISTER} REGISTER_ERROR`;
