+++ src/containers/RegisterPage/index.js
@@ -3,24 +3,15 @@ import { useDispatch, useSelector } from 'react-redux';
 import { Helmet } from 'react-helmet-async';
 import { useIntl } from 'react-intl';
 import { Link } from 'react-router-dom';
-import { useInjectReducer } from 'utils/injectReducer';
-import { useInjectSaga } from 'utils/injectSaga';
 import { makeSelectIsSocialAuthPending } from 'containers/SocialAuth/selectors';
-import { makeSelectIsRegisterPending } from './selectors';
-import { register } from './actions';
+import { makeSelectIsRegisterPending } from 'store/auth/selectors';
+import { register } from 'store/auth/actions';
 import SocialAuth from 'containers/SocialAuth';
 import RegisterForm from './RegisterForm';
-import reducer from './reducer';
-import saga from './saga';
 import messages from './messages';
 import { LOGIN } from 'routes';
 
-const key = 'register';
-
 function RegisterPage() {
-  useInjectReducer({ key, reducer });
-  useInjectSaga({ key, saga });
-
   const dispatch = useDispatch();
   const isRegisterPending = useSelector(makeSelectIsRegisterPending());
   const isSocialAuthPending = useSelector(makeSelectIsSocialAuthPending());
