+++ src/store/auth/actionTypes.js
@@ -12,3 +12,6 @@ export const SESSION_EXPIRED = `${AUTH} SESSION_EXPIRED`;
 export const FORGOT_PASSWORD_REQUEST = `${AUTH} FORGOT_PASSWORD_REQUEST`;
 export const FORGOT_PASSWORD_SUCCESS = `${AUTH} FORGOT_PASSWORD_SUCCESS`;
 export const FORGOT_PASSWORD_ERROR = `${AUTH} FORGOT_PASSWORD_ERROR`;
+export const REGISTER_REQUEST = `${AUTH} REGISTER_REQUEST`;
+export const REGISTER_SUCCESS = `${AUTH} REGISTER_SUCCESS`;
+export const REGISTER_ERROR = `${AUTH} REGISTER_ERROR`;
