+++ src/store/auth/selectors.js
@@ -1,7 +1,11 @@
 import { createSelector } from 'reselect';
 import { initialState } from './reducer';
 import { makeSelectIsLoading } from '../loading/selectors';
-import { LOGIN_REQUEST, FORGOT_PASSWORD_REQUEST } from './actionTypes';
+import {
+  LOGIN_REQUEST,
+  FORGOT_PASSWORD_REQUEST,
+  REGISTER_REQUEST,
+} from './actionTypes';
 
 const selectAuth = (state) => state.auth || initialState;
 
@@ -9,6 +13,8 @@ const makeSelectIsLoginPending = () => makeSelectIsLoading(LOGIN_REQUEST);
 
 const makeSelectIsForgotPasswordPending = () =>
   makeSelectIsLoading(FORGOT_PASSWORD_REQUEST);
+
+const makeSelectIsRegisterPending = () => makeSelectIsLoading(REGISTER_REQUEST);
 
 const makeSelectUser = () =>
   createSelector(selectAuth, (substate) => substate.user);
@@ -25,6 +31,7 @@ const makeSelectToken = () =>
 export {
   makeSelectIsLoginPending,
   makeSelectIsForgotPasswordPending,
+  makeSelectIsRegisterPending,
   makeSelectUser,
   makeSelectError,
   makeSelectIsAuthenticated,
